requirements.txt
absl-py==0.15.0
astor==0.8.1
backports.weakref==1.0.post1
cachetools==3.1.1
certifi==2021.10.8
chardet==4.0.0
click==7.1.2
enum34==1.1.10
Flask==1.1.4
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.3.0
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.41.1
gunicorn==19.10.0
h5py==2.10.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
numpy==1.16.6
oauthlib==3.1.0
opt-einsum==2.3.2
pandas==0.24.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.5
Sastrawi==1.0.1
scipy==1.2.2
six==1.16.0
tensorboard==2.1.0
tensorflow-cpu==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.26.7
Werkzeug==1.0.1
wrapt==1.13.3
